backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
sentence-transformers
pydantic
scikit-learn
numpy
torch
beautifulsoup4
groq 